extract functions into ShowContainer.utils.ts
