+++ tv-shows-app/src/components/features/shows/ShowContainer/ShowContainer.utils.ts
@@ -1,0 +1,13 @@
+import { IReviewList } from "@/typings/review";
+
+export function loadFromLocalStorage(id: string) {
+  const reviewListLoaded = localStorage.getItem("reviews" + id);
+  if (!reviewListLoaded) {
+    return { reviews: [] };
+  }
+  return JSON.parse(reviewListLoaded);
+}
+
+export function saveToLocalStorage(reviewListToSave: IReviewList, id: string) {
+  localStorage.setItem("reviews" + id, JSON.stringify(reviewListToSave));
+}
